% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{condense.data.simple}
\alias{condense.data.simple}
\title{Condenses data from the same times points into mean values}
\usage{
condense.data.simple(data)
}
\arguments{
\item{data}{a data frame of the data to be condensed. The first column contains all of the time points (which may/should be duplicates)}
}
\value{
a matrix of condensed data
}
\description{
Does what it says on the tin
}

