% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{create_polygons}
\alias{create_polygons}
\title{Polygon for prediction intervals}
\usage{
create_polygons(lower, upper)
}
\arguments{
\item{lower}{a LIST of lower bound coordinates}

\item{upper}{a LIST of upper bound coordinates}
}
\value{
a set of coordinates to draw a polygon representing the given prediction intervals
}
\description{
Takes upper and lower bounds as data frames and uses these to produce a polygon for
that this function expects a LIST of bounds (ie. a different polygon for each index in the list),
so a single polygon should come from lower[[1]] and upper[[1]]
}
\seealso{
\code{\link{plot_model_fits}}
}

