% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_prediction_intervals_1.1}
\alias{generate_prediction_intervals_1.1}
\title{Prediction interval generating function}
\usage{
generate_prediction_intervals_1.1(results_MCMC, burnin, param_table, all_data,
  MODEL_FUNCTION)
}
\arguments{
\item{results_MCMC}{a full, valid file path for a .csv file containing the full MCMC chain}

\item{burnin}{numeric value for the number of MCMC chain iterations to be discarded as burn in}

\item{param_table}{a correctly formatted parameter table as loaded by \code{\link{load_param_table}}}

\item{all_data}{a data frame containing all of the original data}

\item{MODEL_FUNCTION}{a pointer to an existing R function that is used to calculate the model trajectories}
}
\value{
a list containing the following items:
\itemize{
\item{model_data: }{a data frame with the original data (ie. fitted) for use by a plotting function}
\item{prediction_data: }{a data frame with the best fitting model's data for use by a plotting function}
\item{lower_prediction_bounds: }{a data frame with the necessary information required to plot the lower prediction bound}
\item{upper_prediction_bounds: }{a data frame with the necessary information required to plot the upper prediction bound}
}
}
\description{
Generates data and prediction intervals that are used to plot model fits with prediction intervals and original data from provided MCMC chain file and parameters. Calls a second helper function, \code{\link{generate_prediction_intervals_1.2}}
}
\seealso{
\code{\link{generate_prediction_intervals_1.2}}
}

