% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_prediction_intervals_1.2}
\alias{generate_prediction_intervals_1.2}
\title{Second tier function for generating prediction intervals from the MCMC chain}
\usage{
generate_prediction_intervals_1.2(mcmc_chains, burnin = NULL, times,
  runs = 100, level = 0.95, smoothing = 0, MODEL_FUNCTION)
}
\arguments{
\item{mcmc_chains}{a matrix containing the whole MCMC chain (ie. parameter values at each iteration)}

\item{burnin}{optional parameter specifying the number of iterations from the MCMC chain that should be discarded as burn in}

\item{times}{a vector of time points to calculate the model trajectory over}

\item{runs}{number of random draws from the multivariate posterior with which to calculate the prediction intervals. Defaults to 100}

\item{level}{desired prediction interval size. Defaults to 0.95 ie. 95 percent prediction intervals}

\item{smoothing}{uses smoothing splines to give smoother prediction intervals where these might be aethsetically spikey. Defaults to 0}

\item{MODEL_FUNCTION}{a pointer to an existing R function that is used to calculate the model trajectories}
}
\value{
a list of the lower and upper prediction lines
}
\description{
This function is called by a wrapper function, \code{\link{generate_prediction_intervals_1.1}} to generate plottable prediction intervals for a model plot.
}
\seealso{
\code{\link{generate_prediction_intervals_1.1}}
}

