% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{generate_prior_data}
\alias{generate_prior_data}
\title{Generates data for prior density plots}
\usage{
generate_prior_data(names, param_table)
}
\arguments{
\item{names}{a vector of parameter names}

\item{the}{parameter table as loaded by \code{\link{load_param_table}}}
}
\value{
a data frame of combined prior densities, with column names "variable (x coord)", "value (y coord)", "param (parameter name)", "chain" (all set to 'prior')
}
\description{
Given a vector of parameter names and the parameter table as loaded by \code{\link{load_param_table}}, returns a data frame that is used to generate density plots of the prior distributions
}

