% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cost_functions.R
\name{likelihood_observation}
\alias{likelihood_observation}
\title{Observation error only function}
\usage{
likelihood_observation(params, data, param_table = NULL,
  optimisation_direction = -1, MODEL_FUNCTION, ...)
}
\description{
Observation error only function
}

