% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{mcmc_all_plots_multi}
\alias{mcmc_all_plots_multi}
\title{Plot all MCMC density and iteration plots}
\usage{
mcmc_all_plots_multi(filename, mcmc_chains, param_table = NULL,
  burnin = NULL)
}
\arguments{
\item{filename}{the overall file name/location to be used, appended with _MCMC_plots.pdf}

\item{mcmc_chains}{the actual MCMC chains to be used to create density plots and iteration plots}

\item{param_table}{OPTIONAL the parameter table as loaded by \code{\link{load_param_table}}, which is then used to create prior density plots}

\item{burnin}{OPTIONAL the burn in period, used to draw a vertical line}
}
\value{
nothing
}
\description{
Saves a pdf file as 'filename'_MCMC_plots.pdf, which contains density and iteration plots for all of the provided parameters.
}
\seealso{
\code{\link{mcmc_iter_multi}}, \code{\link{mcmc_density_multi}}, \code{\link{mcmc_density_single}}, \code{\link{mcmc_iter_single}}
}

