% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{mcmc_density_multi}
\alias{mcmc_density_multi}
\title{Individual MCMC density plots with multiple chains}
\usage{
mcmc_density_multi(name, data, xlims, prior = NULL, best_fits = NULL)
}
\arguments{
\item{name}{the name of the parameter from the data to be plotted}

\item{data}{the whole set of density data. Has a column to separate data by chain number}

\item{xlims}{the x axis bounds for the density plot}

\item{prior}{OPTIONAL set of data to plot the prior distribution}
}
\value{
a single ggplot object of the density plot
}
\description{
Plots the MCMC density plots for the given parameter and data
}
\seealso{
\code{\link{mcmc_all_plots_mcmc}}
}

