% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{mcmc_density_single}
\alias{mcmc_density_single}
\title{Individual MCMC density plots for a single chain}
\usage{
mcmc_density_single(name, data)
}
\arguments{
\item{name}{the name of the parameter from the data to be plotted}

\item{data}{the whole set of density data}
}
\value{
a single ggplot object of the density plot
}
\description{
Plots the MCMC density plots for the given parameter and data
}
\seealso{
\code{\link{mcmc_all_plots_mcmc}}
}

