% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mcmc_wrapper_script.R
\name{MCMC_fit_single}
\alias{MCMC_fit_single}
\title{Bottom level MCMC wrapper function}
\usage{
MCMC_fit_single(data, param_table, iterations = 1000, opt_freq = 50,
  thin = 50, burnin = 100, adaptive_period = 1, nchain = 1,
  popt = 0.44, tmp_filename, LIKELIHOOD_FUNCTION, MODEL_FUNCTION,
  VERBOSE = FALSE, PARALLEL = FALSE)
}
\arguments{
\item{data}{the data over which to calculate likelihoods}

\item{param_table}{table of parameter data as specified in \code{\link{load_param_table}}}

\item{iterations}{number of iterations for the MCMC algorithm}

\item{opt_freq}{how frequently the acceptance rate is adapted. Defaults to 50}

\item{thin}{thinning value for the MCMC chain. Default is 50}

\item{burnin}{the length of the burn in period. Defaults to 100}

\item{adaptive_period}{length of the adaptive period. Defaults to 1}

\item{nchain}{number of chains to run}

\item{popt}{the desired acceptance rate. Defaults to 0.44}

\item{tmp_filename}{the generic file name/location at which to save the MCMC chains. Note that the file extension will be appended to this string}

\item{LIKELIHOOD_FUNCTION}{a valid pointer to an R function which returns a single, log likelihood of the data given the current parameters}

\item{MODEL_FUNCTION}{a valid pointer to an R function which is used to evaluate the model for the current set of parameters}

\item{VERBOSE}{boolean flag for additional output. Defaults to FALSE}

\item{PARALLEL}{OPTIONAL boolean flag indicating if each MCMC chain should be run in parallel using doPar. Defaults to FALSE}
}
\value{
returns a list containing the whole MCMC chains, MCMC summary objects (from coda), the file locations of the MCMC chains, and the maximum likelihood parameters
}
\description{
Calls the \code{\link{run_metropolis_MCMC}} function to carry out the MCMC algorithm. This function handles parallelisation of chains and some basic formatting.
}
\seealso{
\code{\link{run_metropolis_MCMC}}
}

