% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{mcmc_iter_multi}
\alias{mcmc_iter_multi}
\title{Individual MCMC iteration plots with multiple chains}
\usage{
mcmc_iter_multi(name, data, burnin)
}
\arguments{
\item{name}{the name of the parameter to be plotted}

\item{data}{the whole set of data (the MCMC chain). Has a column to separate data by chain number}

\item{burnin}{the length of the burnin period for a vertical line}
}
\value{
a single ggplot object of the iter plot
}
\description{
Plots the MCMC chain for a given parameter and data
}
\seealso{
\code{\link{mcmc_all_plots_mcmc}}
}

