% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc_wrapper_script.R
\name{MCMC_main}
\alias{MCMC_main}
\title{Main MCMC wrapper function}
\usage{
MCMC_main(data_file, param_file, mcmc_param_file, output_file = NULL,
  main_dir = "", group, VERBOSE = FALSE, infection_labels = NULL)
}
\arguments{
\item{data_file}{the correct location of a .csv file containing the data to be fitted. In this case, should have four columns: group, strain, individual, titre and time}

\item{param_file}{the correct location of a .csv file containing information about the parameters that describe the model of interest. See \code{\link{load_param_table}} for further details}

\item{mcmc_param_file}{the correct location of a .csv file containing the parameters for the MCMC fitting. See \code{\link{mcmc_param_check}} for further details}

\item{output_file}{(OPTIONAL) if specified, all standard output will be parsed to this specified file. NOT WELL TESTED, SO RECOMMEND AVOIDING}

\item{main_dir}{top level directory in which to work. Defaults to the current directory}

\item{group}{this is fairly specific to this data. Just specifies which group from the data file should be explored}

\item{VERBOSE}{(OPTIONAL) - if true, will give printed outputs from the program}

\item{infection_labels}{a vector of times and labels to represent infection events. These are used by the model plotting function, and are optional}
}
\value{
1 if an error, 0 otherwise.
}
\description{
This is the main wrapper function for the ferret model data MCMC fits. This will load any necessary data or parameter files and then call the next level MCMC wrapper function, \code{\link{MCMC_fit_1.1}}. Brief description:
Firstly, reads in the parameter files and carries out some checks on these. eg. numeric values where needed, bounds being the right way around etc. Then formats these parameter tables correctly and matches function names with the correct functions. Also melts data to correct format. Creates a set of file names that are used for each sub-group to be fitted (ie. strain). Fairly specific to this problem, also creates an array of infection times
}
\seealso{
\code{\link{MCMC_fit_1.1}}
}

