% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optifix.R
\name{optifix}
\alias{optifix}
\title{Fixed parameter optimisation}
\usage{
optifix(par, fixed, fn, gr = NULL, ..., method = c("Nelder-Mead", "BFGS",
  "CG", "L-BFGS-B", "SANN"), lower = -Inf, upper = Inf, control = list(),
  hessian = FALSE)
}
\arguments{
\item{par}{a vector of the parameters to be optimised}

\item{fixed}{a vector of boolean values corresponding to the values of par. If the corresponding index is TRUE, then the parameter in par will be fixed during optimisation.}

\item{fn}{the cost function to be optimised over, as in optim}

\item{gr}{as in optim}
}
\value{
returns as with optim, but with a few extra variables to summarise the fixed parameters
}
\description{
Optifix is a function originally written by Barry Rowlingson (b.rowlingsno at lancaster.ac.uk) that acts as a wrapper for the optim function, allowing the user to fix any specified parameters during optimisation.
}
\examples{
fit <- optifix(c(1,1),c(TRUE,FALSE),cost_function,data=example_data)
}
\seealso{
\code{\link{optim}}
}

