% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pad_lists}
\alias{pad_lists}
\title{Pads lists of different sizes}
\usage{
pad_lists(my_list)
}
\arguments{
\item{my_list}{list of lists to be padded}
}
\value{
the padded list of lists
}
\description{
When passed a list of lists of different lengths, returns the same list such that each list is the same size, with NAs used to pad the smaller lists. ie. all lists same size as biggest list
}

