% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/checking_functions.R
\name{param_check}
\alias{param_check}
\title{Model parameter table check}
\usage{
param_check(param_table)
}
\arguments{
\item{param_table}{a data frame or matrix containing the parameters and corresponding meta-data.}
}
\value{
returns a string containing an error message, or NULL if no error
}
\description{
Takes the table of parameters to be used for model calculation and MCMC fitting, and returns a string error if any of the inputs are invalid. The table should have the following column names:
\itemize{
\item{"name"}{Name of the parameter, string or character}
\item{"value"}{arbritary numeric value that represents a typical value. This is not that important, it is just used for some upkeep functions}
\item{"lower_bound"}{numeric value for the lower allowable bound of the parameter. This can be -Inf}
\item{"upper_bound"}{numeric value for the upper allowable bound of the parameter. This can be Inf}
\item{"use_logistic"}{boolean value (1/0) indicating whether or not the parameter should be logistic transformed during optimisation, ensuring that the parameter search space is between the specified bounds}
\item{"use_log"}{boolean value (1/0) indicating whether parameter optimisation should be on a log scale. Note that this and use logistic should NOT both be specified}
\item{"step"}{initial step size for the MCMC random walk}
\item{"fixed"}{boolean value (1/0) indicating whether parameter should be fixed during MCMC/optimisation or not}
\item{"nuisance"}{boolean value (1/0) indicating if parameter is a nuisance parameter}
\item{"lower_seed"}{numeric value for the lower bound of the random seed generation}
\item{"upper_seed"}{numeric value for the upper bound of the random seed generation}
\item{"prior_func"}{this should be a string that matches the name of a function that will be used as the prior function for MCMC. This function should be available in the current R environment, or an error will be thrown. See below.}
\item{"prior_args"}{a vector of default arguments for the prior function. For example, this might contain the standard deviation or mean of a normal distribution. Most prior functions have a flag to indicate log probabilities, which should be set to true by this argument}
\item{"log_proposal"}{boolean flag to indicate whether proposals in the MCMC function should be made on a log or linear scale. 1 for log, 0 for linear. If in doubt, set this to 0.}
}
}
\seealso{
\code{\link{mcmc_param_check}}, \code{\link{prior}}, \code{\link{prior_wrapper}}
}

