% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_model_fits}
\alias{plot_model_fits}
\title{Model plotting function}
\usage{
plot_model_fits(model_data, actual_data, group, plot_bounds, lower = NULL,
  upper = NULL, infection_times = NULL, plot_title = NULL)
}
\arguments{
\item{model_data}{a data frame containing the best fit model trajectory. Should be three columns: variable (time), value and strain. Strain is simply a grouping variable and could be anything (eg. age group, location)}

\item{actual_data}{a data frame containing the actual data. Columns are same as model_data}

\item{group}{this is simply used for the file name of the plot. _fit.pdf will be appended}

\item{lower}{OPTIONAL coordinates of the lower prediction bound}

\item{upper}{OPTIONAL coordinates of the upper prediction bound}

\item{infection_times}{OPTIONAL data frame to mark vertical lines representing events. A data frame with columns time and infection, with the time and label respectively}

\item{plot_title}{OPTIONAL plot title}
}
\value{
returns the ggplot object
}
\description{
Provides a neat plot of the provided data, "best fit" line and prediction intervals.
}
\seealso{
\code{\link{restrain_bounds}}, \code{\link{create_ploygons}}
}

