% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{posterior}
\alias{posterior}
\title{Posterior function}
\usage{
posterior(params, param_transform_table, param_table, data, LIKELIHOOD_FUNCTION,
  MODEL_FUNCTION)
}
\arguments{
\item{params}{a vector of parameters to be explored}

\item{param_transform_table}{a matrix of flags and bounds as required by \code{\link{proposalfunction}} for the random walk proposal, for use in the likelihood function}

\item{param_table}{a data table as imported by \code{\link{load_param_table}} to be used by the prior wrapper function in finding the correct prior function and arguments}

\item{data}{the data of interest}

\item{LIKELIHOOD_FUNCTION}{a valid pointer to an R function which returns a single, log likelihood of the data given the current parameters}

\item{MODEL_FUNCTION}{a valid pointer to an R function which is used to evaluate the model for the current set of parameters}
}
\value{
a single value for the posterior (by Bayes rule)
}
\description{
Combines the likelihood and priors to produce a likelihood value for the given parameters against he provided data. This is used as the overall cost function for the MCMC algorithm.
}
\seealso{
\code{\link{prior}}
}

