% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict.titre.fast}
\alias{predict.titre.fast}
\title{Basic ferret model function}
\usage{
\method{predict}{titre.fast}(params, times)
}
\arguments{
\item{params}{the vector of parameters used to solve the model. The first argument should be the lower titre bound first infection time}

\item{times}{a vector of times over which to solve the model. If in doubt, use a sequence from 0 to max time at steps of 1}
}
\value{
a matrix containing the titre values predicted from the given parameters at each time point specified by times
}
\description{
Uses the Ferret boost/wane model in the form y=mx+c to produce a matrix of model results from a given vector of parameters.
}
\seealso{
\code{\link{predict.titre.fast.bounded}}, \code{\link{predict.titre.fast}}, \code{\link{predict_titre_universal_m_2}}, \code{\link{predict_titre_universal_m}}
}

