% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict.titre.OLD}
\alias{predict.titre.OLD}
\title{Old titre prediction function}
\usage{
\method{predict}{titre.OLD}(params, times)
}
\description{
No longer active - this used ODEs rather than y=mx+c which was much too slow
}
\seealso{
\code{\link{predict.titre.fast.bounded}}, \code{\link{predict.titre.fast}}, \code{\link{predict_titre_universal_m_2}}, \code{\link{predict_titre_universal_m}}
}

