% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict_titre_universal_m}
\alias{predict_titre_universal_m}
\title{Ferret model function with a universal waning parameter, m}
\usage{
predict_titre_universal_m(params, times)
}
\arguments{
\item{params}{the vector of parameters used to solve the model. The first argument should be m}

\item{times}{a vector of times over which to solve the model. If in doubt, use a sequence from 0 to max time at steps of 1}
}
\value{
a matrix containing the titre values predicted from the given parameters at each time point specified by times
}
\description{
Uses the Ferret boost/wane model in the form y=mx+c to produce a matrix of model results from a given vector of parameters. This particular version allows a universal waning parameter, m. The first argument should be m
}
\seealso{
\code{\link{predict.titre.fast.bounded}}, \code{\link{predict.titre.fast}}, \code{\link{predict_titre_universal_m_2}}, \code{\link{predict_titre_universal_m}}
}

