% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict_titre_universal_m_2}
\alias{predict_titre_universal_m_2}
\title{Ferret model function with lower bounded titres and a universal waning parameter, m}
\usage{
predict_titre_universal_m_2(params, times)
}
\arguments{
\item{params}{the vector of parameters used to solve the model. The first argument should be m, and the second argument the lower titre bound}

\item{times}{a vector of times over which to solve the model. If in doubt, use a sequence from 0 to max time at steps of 1}
}
\value{
a matrix containing the titre values predicted from the given parameters at each time point specified by times
}
\description{
Uses the Ferret boost/wane model in the form y=mx+c to produce a matrix of model results from a given vector of parameters. This particular version allows the titres to be truncated from below, and a universal waning parameter, m. The first argument should be m, and  the second argument of the parameter vector should provide the lower bound
}
\seealso{
\code{\link{predict.titre.fast.bounded}}, \code{\link{predict.titre.fast}}, \code{\link{predict_titre_universal_m_2}}, \code{\link{predict_titre_universal_m}}
}

