% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rand.params}
\alias{rand.params}
\title{Random parameter generator}
\usage{
rand.params(param_table)
}
\arguments{
\item{param_table}{a data frame of parameters containing columns with the lower and upper seed bounds (titled lower_seed and upper_seed)}
}
\value{
a vector of random parameters
}
\description{
From a table of parameters, generates random values within the given bounds. Idea is to use these as seeds for the MCMC chains
Table should be as imported by \code{\link{load_param_table}}.
}
\seealso{
\code{\link{load_param_table}}
}

