% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_zero_data}
\alias{remove_zero_data}
\title{Remove zero value groups}
\usage{
remove_zero_data(all_data, file_strains)
}
\arguments{
\item{all_data}{a data frame to be processed, with time points, values, and a column "strain" that represent the sub-groups}

\item{file_strains}{a vector of the strain names for the all_data "strain" column}
}
\value{
the same data table and vector of strain names, less those that have all zero values
}
\description{
Removes those sub-groups from the data that have only zero values for all time points, and the corresponding strain name from the strain name list
}

