% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{restrain_bounds}
\alias{restrain_bounds}
\title{Restrains values in a data frame by the given bounds}
\usage{
restrain_bounds(dat, lower, upper)
}
\arguments{
\item{dat}{a data frame to be bounded (with a column y)}

\item{lower}{the lower bound}

\item{upper}{the upper bound}
}
\value{
the same data frame but with bounded y values
}
\description{
Takes a list of data frames (ie. titre data) and makes sure that all y values are between lower and upper
}

