% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{scaletuning1}
\alias{scaletuning1}
\title{Scale step sizes}
\usage{
scaletuning1(step, popt, pcur)
}
\arguments{
\item{step}{the current step size}

\item{popt}{the desired acceptance rate}

\item{the}{current acceptance rate}
}
\value{
the scaled step size
}
\description{
Scales the given step size (between 0 and 1) based on the current acceptance rate to get closed to the desired acceptance rate
}

