% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_functions.R
\name{titre_model}
\alias{titre_model}
\title{ODE for ferret model}
\usage{
titre_model(t, x, params)
}
\arguments{
\item{t}{vector of time points over which dY should be calculated}

\item{x}{nothing - just required by deSolve I think...}

\item{params}{the vector of parameters for the model}
}
\description{
ODE for Boosting and Waning model, used by the ode solver, deSolve.
}

