% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{transform_params_logistic}
\alias{transform_params_logistic}
\title{Parameter logistic/exp transform function}
\usage{
transform_params_logistic(params, param_table)
}
\arguments{
\item{params}{vector of parameters to be transformed}

\item{param_table}{a matrix of boolean flags and bounds}
}
\value{
a vector of the transformed parameters
}
\description{
Takes a list of parameters and table of parameter bounds. Idea is to convert all parameters from log/specified logit scale to normal scale (-Inf to +Inf).
Table should have four columns: use_logistic, use_log, lower_bound, and upper_bound.
}
\seealso{
\code{\link{param_check}}, \code{\link{transform_params_logit}}
}

