% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{transform_params_logisticOLD}
\alias{transform_params_logisticOLD}
\title{OLD transform parameters to logistic}
\usage{
transform_params_logisticOLD(params, param_table)
}
\description{
Takes a list of parameters and table of parameter bounds. Idea is to convert all parameters from log/specified logit scale to normal scale (-Inf to +Inf).
}
\seealso{
\code{\link{transform_params_logistic}}
}

