% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{transform_params_logit}
\alias{transform_params_logit}
\title{Parameter logit/log transform function}
\usage{
transform_params_logit(params, param_table)
}
\arguments{
\item{params}{vector of parameters to be transformed}

\item{param_table}{a matrix of boolean flags and bounds}
}
\value{
a vector of the transformed parameters
}
\description{
Takes a vector of parameters and a table of parameter bounds, transformation instructions etc. Idea is to convert all parameters to log/specified logit scale.
Table should have four columns: use_logistic, use_log, lower_bound, and upper_bound.
}
\seealso{
\code{\link{param_check}}, \code{\link{transform_params_logistic}}
}

