% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{transform_params_logitOLD}
\alias{transform_params_logitOLD}
\title{OLD transform parameters to logit}
\usage{
transform_params_logitOLD(params, param_table)
}
\description{
Takes a list of parameters and a table of parameter bounds, transformation instructions etc. Idea is to convert all parameters to log/specified logit scale.
}
\seealso{
\code{\link{transform_params_logit}}
}

